Adding jac_mpi to jac1 application directory

+++ apps/c/jac1/jac_kernels.cpp
@@ -10,7 +10,7 @@ void __syncthreads(){}
                             
 // global constants         
                             
-extern float alpha;         
+extern double alpha;         
                             
 // user kernel files
